fix:refactored pydantic extension registration (#1152)
* refactored pydantic extension registration

* delete data_quality/pydantic_validators, content moved to plugins/pydantic_extensions

* added pydantic 2.0 requirements to docstrings

---------
