--- tests/integrations/pydantic/requirements.txt
@@ -1,1 +1,0 @@
-# Additional requirements on top of hamilton...pydantic
